+++ components/ChatBot.tsx
@@ -1,5 +1,17 @@
-import { Chat } from "./chatbot/Chat";
-import { useEffect, useRef, useState } from "react";
+'use client'
+
+import Image from 'next/image'
+import React, { useEffect, useRef, useState } from 'react'
+import { FaRegFaceGrin } from 'react-icons/fa6'
+import { ImBubbles2 } from 'react-icons/im'
+import { IoClose } from 'react-icons/io5'
+import { MdAttachFile } from 'react-icons/md'
+import { RiVoiceprintLine } from 'react-icons/ri'
+import avatar from '../assets/default_avatar.png'
+import crisp from '../assets/crisp.svg'
+import logo from '../assets/logo.svg'
+import woman from '../assets/woman.jpg'
+import { Chat } from './chatbot/Chat'
 
 interface Message {
   role: Role;
@@ -8,7 +20,8 @@ interface Message {
 
 type Role = "assistant" | "user";
 
-export default function Chatbot() {
+export default function Popup () {
+  const [isOpen, setIsOpen] = useState(true)
   const [messages, setMessages] = useState<Message[]>([]);
   const [loading, setLoading] = useState<boolean>(false);
 
@@ -110,19 +123,94 @@ export default function Chatbot() {
     ]);
   }, []);
 
+  const handleOpen = () => {
+    setIsOpen(prev => !prev)
+  }
+
+  const images = [
+    avatar,
+    logo,
+    woman,
+  ]
+
   return (
-    <>
-        <div className="absolute bottom-10 right-0 flex-1 overflow-auto sm:px-10 pb-4 sm:pb-10">
-          <div className="max-w-[500px] mx-auto mt-4 sm:mt-12">
-            <Chat
-              messages={messages}
-              loading={loading}
-              onSend={handleSend}
-              onReset={handleReset}
-            />
-            <div ref={messagesEndRef} />
+    <div className='fixed overflow-hidden w-screen flex flex-col-reverse z-20 gap-3 p-4 items-end justify-center bottom-0 right-0 md:bottom-1 md:px-4 md:w-2/4 md:py-2 lg:w-[28%] h-fit'>
+      <button onClick={handleOpen} className='bg-black text-white h-12 w-12 rounded-full flex items-center justify-center drop-shadow-xl cursor-pointer'>
+        {
+        isOpen
+          ? <IoClose className='h-full w-auto p-2.5' />
+          : <ImBubbles2 className='h-full w-auto p-3.5' />
+      }
+      </button>
+      <div className={`grow shadow-xl rounded-xl w-full flex-col h-[90vh] pt-4 md:h-[80vh] md:pt-0 ${isOpen ? 'flex' : 'hidden'}`}>
+        <div className='bg-zinc-800 rounded-t-xl p-2 pb-3 flex flex-col items-center gap-2'>
+          <div className='text-white rounded-xl px-7 py-1.5 text-xs bg-neutral-700 flex items-center gap-2'>
+            <span><ImBubbles2 /></span>
+            <span className='font-bold text-xs'>Chat</span>
+          </div>
+          <ul className='flex'>
+            {
+            images.map((image, i) => (
+              <li key={i} className={`h-9 w-9 relative rounded-full overflow-hidden border-2 border-zinc-800 ${i > 0 && '-ms-1.5'}`}>
+                <Image src={image} alt='Avatar' className='h-full w-auto' fill />
+              </li>
+            ))
+          }
+          </ul>
+          <div className='text-center text-xs'>
+            <div className='text-white font-bold mb-0.5'>Questions? Chat with us!</div>
+            <p className='text-slate-300'>Was last active 11 hours ago</p>
           </div>
         </div>
-    </>
-  );
+        <div className='bg-white text-black rounded-b-lg p-2 grow flex flex-col'>
+          <div className='grow py-2 border-b'>
+            <ul>
+              <li className='flex gap-2'>
+                <div className='h-7 w-7 rounded-full overflow-hidden flex items-center justify-center relative'>
+                  <Image src={logo} alt='Avatar' className='h-full w-auto' fill />
+                </div>
+                <div className='flex flex-col gap-1'>
+                  <div className='text-slate-400 text-xs'>Writier</div>
+                  <div className='px-2.5 py-2 bg-zinc-800 text-white rounded-md text-xs'>How can we help with Writier?</div>
+                </div>
+              </li>
+            </ul>
+          </div>
+          {/* <div className="absolute bottom-10 right-0 flex-1 overflow-auto sm:px-10 pb-4 sm:pb-10">
+            <div className="max-w-[500px] mx-auto mt-4 sm:mt-12">
+              <Chat
+                messages={messages}
+                loading={loading}
+                onSend={handleSend}
+                onReset={handleReset}
+              />
+              <div ref={messagesEndRef} />
+            </div>
+          </div> */}
+          <div className='p-2 flex flex-col gap-4'>
+            <input type='text' placeholder='Compose your message' className='py-2 text-xs outline-none' />
+            <div className='flex items-center justify-between'>
+              <div className='flex items-center text-slate-600 gap-2'>
+                <button className='h-3.5 w-3.5'>
+                  <FaRegFaceGrin className='h-full w-auto' />
+                </button>
+                <button className='h-3.5 w-3.5'>
+                  <MdAttachFile className='h-full w-auto' />
+                </button>
+                <button className='h-3.5 w-3.5'>
+                  <RiVoiceprintLine className='h-full w-auto' />
+                </button>
+              </div>
+              <div className='flex items-center gap-1'>
+                <div className='text-slate-400 text-xs'>We run on</div>
+                <div className='h-3 w-10 relative'>
+                  <Image src={crisp} alt='Crisp logo' className='h-full w-auto' fill />
+                </div>
+              </div>
+            </div>
+          </div>
+        </div>
+      </div>
+    </div>
+  )
 }
